# core/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "All the endpoints for Reminderse API. Hosted on https://api.reminderse.com",
    "version": "1.0.0",
    "title": "Reminderse API",
    "contact": {
      "email": "sherzodnimatullo@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "tags": [
    {
      "name": "Users",
      "description": "Endpoints for managing users"
    },
    {
      "name": "Text Entries",
      "description": "Endpoints for managing text entries"
    },
    {
      "name": "Link Entries",
      "description": "Endpoints for managing link entries"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Authenticate a user",
        "requestBody": {
          "description": "Login Request",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginRequest"
              }
            }
          }
        },
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/successfulLoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect log in credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": ["Users"],
        "summary": "Register a new user",
        "requestBody": {
          "description": "Register request",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/registerRequest"
              }
            }
          }
        },
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful signup",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          },
          "400": {
            "description": "Username or email already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          }
        }
      }
    },
    "/change/username": {
      "put": {
        "tags": ["Users"],
        "summary": "Change username for currently logged in user",
        "requestBody": {
          "description": "Change username request",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/changeUsernameRequest"
              }
            }
          }
        },
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful change of username",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          },
          "400": {
            "description": "Username already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          }
        }
      }
    },
    "/change/email": {
      "put": {
        "tags": ["Users"],
        "summary": "Change email for currently logged in user",
        "requestBody": {
          "description": "Change email request",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/changeEmailRequest"
              }
            }
          }
        },
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful change of email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          },
          "400": {
            "description": "Email already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          }
        }
      }
    },
    "/change/password": {
      "put": {
        "tags": ["Users"],
        "summary": "Change password for currently logged in user",
        "requestBody": {
          "description": "Change password request",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/changePasswordRequest"
              }
            }
          }
        },
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful change of password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          }
        }
      }
    },
    "/unsubscribe": {
      "delete": {
        "tags": ["Users"],
        "summary": "Unsubscribe and delete the account.",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful unsubscribe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          }
        }
      }
    },
    "/current-user": {
      "get": {
        "tags": ["Users"],
        "summary": "Get currently logged in user",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/successfulLoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/link/": {
      "get": {
        "tags": ["Link Entries"],
        "summary": "Get all links",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/links"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Link Entries"],
        "summary": "Create a new link",
        "requestBody": {
          "description": "Create link request",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createLinkRequest"
              }
            }
          }
        },
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "Successful creation of link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/message"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "message": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "loginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "registerRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "successfulLoginResponse": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "changeUsernameRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          }
        }
      },
      "changeEmailRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        }
      },
      "changePasswordRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "format": "password"
          },
          "newPassword": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "links": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["id", "title", "url", "days"],
              "properties": {
                "id": {
                  "type": "string"
                },
                "entry_title": {
                  "type": "string"
                },
                "url": {
                  "type": "string"
                },
                "days": {
                  "type": "integer"
                },
                "category": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "createLinkRequest": {
        "type": "object",
        "required": ["title", "url", "days"],
        "properties": {
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "days": {
            "type": "integer"
          },
          "category": {
            "type": "string"
          }
        }
      }
    }
  }
}